requirements.txt
fastapi~=0.115.4
uvicorn~=0.32.0
SQLAlchemy~=2.0.36
aiomysql~=0.2.0
python-decouple~=3.8
pydantic~=2.9.2
typing_extensions~=4.12.2
bcrypt~=4.2.1
jose~=1.0.0
python-jose~=3.3.0
python-multipart~=0.0.19
cryptography~=44.0.0
email-validator
Jinja2